Delay buffer is roughly working
The "manual" sim shows that it is taking in and outputting signals now, and can take positive or negative extra delays. However, not all the edge cases have been covered. It seems there's still a bug when totalDelay = 0.

I think I might want to eventually remake this module as an FSM so I can put it in a boot-up mode which pre-loads the buffer with zeros.