Async FIFO module + testbench isn't working
The write side is working (writes four packets then full flag goes high) but the read side is not because the empty flag never goes low.

Debug steps:
* Switching polarity on reset did nothing
* Tying read enable always high did nothing
* Found logic inside the FIFO IP module which suggests that it's properly writing four times and successfully reads from the EBR *once*

Gemini thinks that something is causing the read pointer to not synchronize from writePtr -> grayPtr -> readPtr, which is causing the empty flag to stay high and therefore the data read from the EBR is never being output. It suspects there may be something wrong with the reset.